requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
cycler==0.10.0
decorator==4.3.0
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
joblib==0.11
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
networkx==2.1
numpy==1.14.3
opencv-python==3.4.1.15
pandas==0.23.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
